+++ src/Models/Dud.php
@@ -1,9 +1,0 @@
-<?php
-
-namespace Laraquick\Models;
-
-use Illuminate\Database\Eloquent\Model;
-
-class Dud extends Model
-{
-}