+++ src/Models/Traits/Helper.php
@@ -55,34 +55,6 @@ trait Helper
         return $query;
     }
 
-    /**
-     * Determine if the given relation is loaded.
-     *
-     * @param  string  $key
-     * @return bool
-     */
-    public function relationLoaded($key)
-    {
-        if (!Str::contains($key, '.')) {
-            return isset($this->relations[$key]);
-        }
-
-        $firstKey = Str::before($key, '.');
-        $otherKeys = Str::after($key, '.');
-
-        if (!isset($this->relations[$firstKey])) {
-            return false;
-        }
-
-        $relation = $this->relations[$firstKey];
-
-        if ($relation instanceof EloquentCollection) {
-            return $relation->contains(fn (Model $model) => $model->relationLoaded($otherKeys));
-        }
-
-        return optional($relation)->relationLoaded($otherKeys) ?? false;
-    }
-
     public function toArray(): array
     {
         $fillable = $this->fillable ?? [];