+++ CHANGELOG.md
@@ -7,10 +7,15 @@
 ### Updated
 
 - Fixed toArray issue with snake/camel case relationships.
+- Adds method `getMailMessage` to Notifications\Send.
 
 ### Removed
 
 - Removed Helper/Http. Laravel now has an Http facade.
+- Removed Events/WebSocket. Use [Laravel WebSockets](https://beyondco.de/docs/laravel-websockets/getting-started/introduction).
+- Removed Helper::relationLoaded method. Laravel now has a method for that.
+
+## Missed out on a couple of version. Will try to this file up-to-date going forward.
 
 ## 3.8.0