+++ CHANGELOG.md
@@ -25,6 +25,8 @@
 - Removed Events/WebSocket. Use [Laravel WebSockets](https://beyondco.de/docs/laravel-websockets/getting-started/introduction).
 - Removed Helper::relationLoaded method. Laravel now has a method for that.
 - Removed Helper/Excel.
+- Removed [d-scribe/laravel-db-command](https://github.com/ezra-obiwale/laravel-db-command) as a dependency.
+- Removed JWT references entirely
 
 ## Missed out on a couple of version. Will try to this file up-to-date going forward.