+++ CHANGELOG.md
@@ -18,6 +18,7 @@
 - Swaps [d-scribe/laravel-apidoc-generator](https://github.com/d-scribe/laravel-apidoc-generator) with [Scribe](https://scribe.knuckles.wtf/laravel).
 - Changes 'all' to `-1` on index action to load all.
 - Fixes storing test responses with 204 content
+- Renames config tag from just 'config' to 'laraquick-config'
 
 ### Removed